requirements.txt
# 核心框架
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
langchain-openai>=0.0.5

# 向量数据库
pymilvus>=2.3.0
faiss-cpu>=1.7.4

# 可视化
streamlit>=1.28.0
plotly>=5.17.0

# 工具库
python-dotenv>=1.0.0
pyyaml>=6.0
tiktoken>=0.5.0
openai>=1.10.0

# 日志和工具
colorama>=0.4.6
rich>=13.7.0

# 可选：DeepSeek API
# 如果使用 DeepSeek，需要安装对应的 SDK
# dashscope>=1.17.0

